Refactor FeaturePreview location and update copy
