+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 // src/app/page.tsx
 
 import ContinuePlanningBanner from '@/features/home/components/ContinuePlanningBanner';
-import FeaturePreview from '@/features/home/components/feature-preview/FeaturePreview';
+import FeaturePreview from '@/features/home/components/FeaturePreview';
 import FinalCta from '@/features/home/components/FinalCta';
 import Hero from '@/features/home/components/Hero';
 import HomeFooter from '@/features/home/ui/HomeFooter';
